--- bindata/generate.go
@@ -1,3 +1,0 @@
-package bindata
-
-//go:generate go-bindata -pkg bindata -modtime 1587356420 -o bindata.go files/ ../examples/dnclient.conf.example
