+++ pkg/bindata/README.md
