+++ pkg/bindata/files/favicon.ico
