+++ pkg/bindata/files/macos.png
